Create a default for <leader> to display in RiVimStatusLine
This function is trying to get the value of <Leader> in a variable
to display to the user in the status line.  Seems that unless set
by user in vimrc, Vim doesn't register g:mapleader.  Assume '\' as
that is the default (on Linux and other *nix's; I can't speak for Mac)